Sync core: 2909: Fix for quest errors 2911: Castle/fort/clanhall bonus for regeneration now wokrking properly 2913: Two little fixes on BitSetIDFactory synched operations 2917: Reflect damage fix, thanks _DS_ 2919: Added sorted wh for clan hall manager and castle 2920: Pet manager html path fix, thanks Charus.
git-svn-id: http://svn.l2jfree.com/l2jfree@5734 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/skills/Formulas.java
@@ -1090,7 +1090,7 @@ public final class Formulas
 					ClanHall clansHall = ClanHallManager.getInstance().getClanHallById(clanHallIndex);
 					if (clansHall != null)
 						if (clansHall.getFunction(ClanHall.FUNC_RESTORE_HP) != null)
-							hpRegenMultiplier *= 1 + clansHall.getFunction(ClanHall.FUNC_RESTORE_HP).getLvl() / 100;
+							hpRegenMultiplier *= 1 + (double) clansHall.getFunction(ClanHall.FUNC_RESTORE_HP).getLvl() / 100;
 				}
 			}
 
@@ -1106,7 +1106,7 @@ public final class Formulas
 					Castle castle = CastleManager.getInstance().getCastleById(castleIndex);
 					if (castle != null)
 						if (castle.getFunction(Castle.FUNC_RESTORE_HP) != null)
-							hpRegenMultiplier *= 1 + castle.getFunction(Castle.FUNC_RESTORE_HP).getLvl() / 100;
+							hpRegenMultiplier *= 1 + (double) castle.getFunction(Castle.FUNC_RESTORE_HP).getLvl() / 100;
 				}
 			}
 
@@ -1118,7 +1118,7 @@ public final class Formulas
 					Fort fort = FortManager.getInstance().getFortById(fortIndex);
 					if (fort != null)
 						if (fort.getFunction(Fort.FUNC_RESTORE_HP) != null)
-							hpRegenMultiplier *= 1 + fort.getFunction(Fort.FUNC_RESTORE_HP).getLvl() / 100;
+							hpRegenMultiplier *= 1 + (double) fort.getFunction(Fort.FUNC_RESTORE_HP).getLvl() / 100;
 				}
 			}
 
@@ -1201,7 +1201,7 @@ public final class Formulas
 					Castle castle = CastleManager.getInstance().getCastleById(castleIndex);
 					if (castle != null)
 						if (castle.getFunction(Castle.FUNC_RESTORE_MP) != null)
-							mpRegenMultiplier *= 1 + castle.getFunction(Castle.FUNC_RESTORE_MP).getLvl() / 100;
+							mpRegenMultiplier *= 1 + (double) castle.getFunction(Castle.FUNC_RESTORE_MP).getLvl() / 100;
 				}
 			}
 
@@ -1213,7 +1213,7 @@ public final class Formulas
 					Fort fort = FortManager.getInstance().getFortById(fortIndex);
 					if (fort != null)
 						if (fort.getFunction(Fort.FUNC_RESTORE_MP) != null)
-							mpRegenMultiplier *= 1 + fort.getFunction(Fort.FUNC_RESTORE_MP).getLvl() / 100;
+							mpRegenMultiplier *= 1 + (double) fort.getFunction(Fort.FUNC_RESTORE_MP).getLvl() / 100;
 				}
 			}
 
@@ -1225,7 +1225,7 @@ public final class Formulas
 					ClanHall clansHall = ClanHallManager.getInstance().getClanHallById(clanHallIndex);
 					if (clansHall != null)
 						if (clansHall.getFunction(ClanHall.FUNC_RESTORE_MP) != null)
-							mpRegenMultiplier *= 1 + clansHall.getFunction(ClanHall.FUNC_RESTORE_MP).getLvl() / 100;
+							mpRegenMultiplier *= 1 + (double) clansHall.getFunction(ClanHall.FUNC_RESTORE_MP).getLvl() / 100;
 				}
 			}
